Crate frame_support_test_compile_pass

This crate tests that construct_runtime! expands the pallet parts correctly even when frame-support is renamed in Cargo.toml

Structs

BlockHashCount

GenesisConfig

Origin

The runtime origin type represanting the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

SS58Prefix

Version

Enums

Call

Event

OriginCaller

Constants

VERSION

Type Definitions

AccountId

AllModulesDeprecated

All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystemDeprecated

All modules included in the runtime as a nested tuple of types.

AllPallets

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

Block

BlockNumber

Header

Index

Signature

System

SystemConfig

UncheckedExtrinsic