Crate frame_support_test_compile_pass

This crate tests that construct_runtime! expands the pallet parts correctly even when frame-support is renamed in Cargo.toml

Structs

GenesisConfig

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeOrigin

The runtime origin type representing the origin of a call.

Version

Enums

OriginCaller

RuntimeCall

RuntimeEvent

Constants

VERSION

Type Definitions

AccountId

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Block

BlockNumber

Header

Index

Signature

System

SystemConfig

UncheckedExtrinsic